% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_to_database.R
\name{analysisToDatabase}
\alias{analysisToDatabase}
\title{analysis to Database}
\usage{
analysisToDatabase(workshop_dir)
}
\arguments{
\item{workshop_dir}{dir of workshop}
}
\description{
analysis to Database
}
\details{
This function will analysis the prediction of distribution, give many optimal threshold into database
}

