% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apply_binary_method.R
\name{apply_threshold_binary_distribution}
\alias{apply_threshold_binary_distribution}
\title{apply_threshold_binary_distribution}
\usage{
apply_threshold_binary_distribution(workshop_dir, algorithm = TRUE,
  threshold_method = TRUE, select_method = TRUE, standard = TRUE)
}
\arguments{
\item{workshop_dir}{Directory of workshop}

\item{algorithm}{Which modelling algorithm(s) resutl user want to binary distribution.
Default is all that used in SDMengine modelling.}

\item{threshold_method}{How to choose threshold. Defalut is all that SDManalysis can provide}

\item{select_method}{How to choose method to choose threshold}

\item{standard}{The cites which will been used in select_method}
}
\description{
apply_threshold_binary_distribution
}

