% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_database_to_data_frame.R
\name{load_database_to_data_frame}
\alias{load_database_to_data_frame}
\title{load_database_to_data_frame}
\usage{
load_database_to_data_frame(workshop_dir)
}
\arguments{
\item{workshop_dir}{dir of workshop}
}
\value{
data.frame of result
}
\description{
load_database_to_data_frame
}

