% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report_analysis.R
\name{report_analysis}
\alias{report_analysis}
\title{report calculate result}
\usage{
report_analysis(workshop_dir, csv_file)
}
\arguments{
\item{workshop_dir}{Directory of workshop}

\item{csv_file}{The file which will contain the output result (CSV format)}
}
\description{
report calculate result
}

