% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/signal_instance.R
\name{signal_instance}
\alias{signal_instance}
\title{Signal CPU instance}
\usage{
signal_instance(parameter, workshop_dir, configure)
}
\arguments{
\item{parameter}{run parameter}

\item{workshop_dir}{Directory of workshop}

\item{configure}{Configure object}
}
\description{
Signal CPU instance
}

